@@ -11,10 +11,10 @@ description: A new Flutter application.
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.1+5
+version: 2.3.0
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.14.0"
 
 dependencies:
   flutter:
@@ -23,11 +23,12 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.3
-  shared_preferences: ^0.5.7+2
-  assets_audio_player: ^2.0.13+1
-  rate_my_app: ^0.7.2
-  health_kit_reporter: ^1.0.6
+  cupertino_icons: ^1.0.6
+  shared_preferences: ^2.2.1
+  assets_audio_player: ^3.1.1
+  rate_my_app: ^2.0.0
+  # health_kit_reporter: ^2.1.1
+  file: ^6.1.4
 
 dev_dependencies:
   flutter_test:
